// src/views/system/role/utils/types.ts
// 虽然字段很少 但是抽离出来 后续有扩展字段需求就很方便了

interface FormItemProps {
  /** 角色 ID（编辑时） */
  id?: number;
  /** 角色名称 */
  name: string;
  /** 备注 */
  remark: string;
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };